feat: refactor server plugin type
