+++ packages/server/core/src/types/plugins/index.ts
@@ -1,0 +1,13 @@
+export * from './new';
+export * from './old';
+export type {
+  ServerConfig,
+  CacheConfig,
+  OnFallback,
+  FallbackReason,
+  GetRenderHandlerOptions,
+  NodeServer,
+  NodeRequest,
+  NodeResponse,
+  FileChangeEvent,
+} from './base';
